Speed improvements, updating while processing
Switched use of innerHTML to adding child elements to the DOM (reported to be faster on StackOverflow).

Also added a timeout so that the page could add each row of statistics as it is available. Requires waiting 300ms in order to do next row, which adds some pretty significant lag, but seems unavoidable, as <300 doens't update the page until the entire function is done.