Changed sort, modified table, possible speed increase
Changed sort to prioritize shorter genomes when fitness >0 and genome.length > 45. This may be a bad idea. See code comments.

Modified table to add a border between generations and to include a table header

Changed font to monospace to make it easier to see genome

Reduced calls to spaceEvery4(). This might improve the speed of the runs, but it will prevent the genome from ever saying something like AB CDEB 123A. That could happen before because the truncating section might truncate from the middle of a group of 4 to the end of a group of 4, causing a random group of 2 (or 1 or 3). Regardless, it wasn't really necessary to be using spaceEvery4 all the time, except when printing the genome out for the table, so we might as well cut it and reduce calls to it. This also reduces calls to join and split.