Added Graphs, attempted punctuated equilibrium
Wrote a JS graphing library, which will show average fitness, max fitness, and genome length. Currently limited to doing the first 800 generations.

Also found a bug, which on very rare occasions will cause max fitness to not be the actual max fitness. Not sure where this is coming from.

The switch to "all other things being equal, prefer the shorter genome" resulted in less evolution. Switched it so that it will not care about shorter genome for the first 1000 generations, but then will spend 200 generations where if possible it will make it shorter. I figure that this way it will attempt to cut out random moves and maybe we'll get lucky by cutting out a chunk where a piece hits another due to a frivilous move.