Modified Graphs, touched up CSS
Simplified CSS some (reset issue)

Put Graph into a div that will automatically add a scroll bar so it can grow as far to the side as we'd like. In the future may want to add a autoscroll option for that div, so that we can keep moving to the right as it goes.